Improve readability and reliability in RozetkaPage
Updated the CSS selector in the 'openCart()' method to a more readable one. The previous selector (.catalog-grid__cell:nth-child(1) .goods-tile__prices:nth-child(7) svg:nth-child(1)) was replaced with a less complicated and more readable version (li[class*='item--cart']). These changes make the code cleaner and easier to maintain. Additionally, it enhances the reliability of the tests by selecting the element in a more efficient and less error-prone manner.