Optimize browser window management and update Selenium version
This commit removes the argument '--start-maximized' from the ChromeOptions in RozetkaTestSuite.java and instead moves the window maximizing method to the PageFactory constructor in PageFactory.java. This change will ensure that the browser window is maximized whenever a new PageFactory object is created, thus improving the management of the browser window throughout the runtime. Also, updated Selenium version to 4.12.0 in pom.xml for better performance and added features.