Improve product details retrieval and Chrome driver setup
In RozetkaPage.java, added an explicit wait to ensure the visibility of the product name element in the cart before attempting to retrieve its text. This modification will result in more robust code, reducing the risk of NoSuchElementException. Also, reduced complexity in Chrome driver setup in RozetkaTestSuite.java by removing the start-maximized option, since it is not essential for the test, streamlining the setup process. Further, the wait time for the text verification in test suite is now tuned to 'Также рекомендуем' for better accuracy.